Remove the spawn provider; add ignited. Split storage and manifeststorage
