+++ pkg/providers/storage/storage.go
@@ -1,4 +1,4 @@
-package providers
+package storage
 
 import (
 	log "github.com/sirupsen/logrus"
@@ -7,7 +7,6 @@ import (
 	"github.com/weaveworks/ignite/pkg/providers"
 	"github.com/weaveworks/ignite/pkg/storage"
 	"github.com/weaveworks/ignite/pkg/storage/cache"
-	"github.com/weaveworks/ignite/pkg/storage/manifest"
 )
 
 func SetGenericStorage() error {
@@ -15,15 +14,5 @@ func SetGenericStorage() error {
 	providers.Storage = cache.NewCache(
 		storage.NewGenericStorage(
 			storage.NewDefaultRawStorage(constants.DATA_DIR), scheme.Serializer))
-	return nil
-}
-
-func SetManifestStorage() error {
-	log.Trace("Initializing the ManifestStorage provider...")
-	ms, err := manifest.NewManifestStorage("/etc/firecracker/manifests")
-	if err != nil {
-		return err
-	}
-	providers.Storage = cache.NewCache(ms)
 	return nil
 }
